requirements.txt
pydantic~=2.10.6
SQLAlchemy~=2.0.38
python-dotenv~=1.0.1
fastapi~=0.115.11
stripe~=11.6.0
PyJWT~=2.10.1
redis~=5.2.1
passlib~=1.7.4
aiosmtplib~=4.0.0
httpx~=0.28.1
requests~=2.32.3
alembic~=1.14.1